backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
boto3
langchain
langchain-aws
pydantic
numpy
pandas 